--- Added --- - Hit particle - Readded WaveSpawner

+++ Neon Archery 2D/Assets/Scripts/ScriptablePath.cs
@@ -10,6 +10,6 @@ public class ScriptablePath : ScriptableObject
 	public float speed;
 
 	// Target Spawning
-	public float waveSize;
+	public int waveSize;
 	public float waveSpacing;
 }
